# server-go/docs/swagger.yaml
basePath: /api
definitions:
  handlers.CategoryInput:
    properties:
      name:
        type: string
      parentId:
        type: integer
      slug:
        type: string
    type: object
...
  models.Category:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/models.Category'
      parentId:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  models.CreateAddressInput:
    properties:
      addressLine1:
        type: string
      addressLine2:
        type: string
      city:
        type: string
      fullName:
        type: string
      isDefault:
        type: boolean
      phone:
        type: string
      state:
        type: string
      zipCode:
        type: string
    required:
    - addressLine1
    - city
    - fullName
    - phone
    - state
    - zipCode
    type: object
...
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      images:
        type: string
      isActive:
        type: boolean
      isFeatured:
        type: boolean
      name:
        type: string
      price:
        type: number
      slug:
        type: string
      stock:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.UpdateAddressInput:
    properties:
      addressLine1:
        type: string
      addressLine2:
        type: string
      city:
        type: string
      fullName:
        type: string
      phone:
        type: string
      state:
        type: string
      zipCode:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@bycigar.com
    name: API Support
  description: BYCIGAR 电商后台 API 接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BYCIGAR API
  version: "1.0"
paths:
  /addresses:
    get:
      consumes:
      - application/json
      description: 获取当前用户的收货地址列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取地址列表
      tags:
      - addresses
    post:
      consumes:
      - application/json
      description: 创建新的收货地址
      parameters:
      - description: 地址信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateAddressInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建地址
      tags:
      - addresses
  /addresses/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的收货地址
      parameters:
      - description: 地址ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除地址
      tags:
      - addresses
    put:
      consumes:
      - application/json
      description: 更新指定的收货地址
      parameters:
      - description: 地址ID
        in: path
        name: id
        required: true
        type: integer
      - description: 地址信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateAddressInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新地址
      tags:
      - addresses
  /addresses/{id}/default:
    put:
      consumes:
      - application/json
      description: 将指定地址设为默认收货地址
      parameters:
      - description: 地址ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 设置默认地址
      tags:
      - addresses
...
  /admin/categories:
    get:
      consumes:
      - application/json
      description: 获取所有分类列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
      security:
      - BearerAuth: []
      summary: 获取分类列表(管理端)
      tags:
      - admin-categories
    post:
      consumes:
      - application/json
      description: 创建新的商品分类
      parameters:
      - description: 分类信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.CategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建分类
      tags:
      - admin-categories
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定分类
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除分类
      tags:
      - admin-categories
    put:
      consumes:
      - application/json
      description: 更新指定分类信息
      parameters:
      - description: 分类ID
        in: path
        name: id
        required: true
        type: integer
      - description: 分类信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.CategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新分类
      tags:
      - admin-categories
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"